Add a horrible hack.
Okay, so when we have:

		INCLUDE	"file.a"

We want that line to be included in the listing file *before* the contents
of the file being included.  So I have included a hack just for this
situation, involving a flag being passed around.  It's a hack, but according
to James Hague <https://prog21.dadgum.com/27.html> it's okay.