Bug fix---phase errors should not happen now
The phase errors happened because of forward references where pass 1 made
the assumption of a 16-bit address/offset, and pass 2 finally made the
decision of an 8-bit address/offset but didn't recall the previous
assumption.  So fix the base error, add a few warnings about it, and fix the
phase error message since that cause shouldn't happend again.