Rename trigger to Assert.
I originally tried using *assert* but in some cases, that confused the
compiler because of the assert() macro.  So I changed the name to trigger.
But that's not quite the right name, and it gives off the wrong conntation,
so let's just rename things to Assert, with a capital 'A'.