Bug fix---adjust when .OPT TEST directives apply
When using the test backend, the .OPT directive can appear either in a .TEST
directive, or outside.  When outside, execute the option when read;
otherwise, execute the directive when the test is run.  Outside a .TEST
directive, the .OPT can be used to set up options that apply to all the
tests prior to the run.  Inside a .TEST directive, they only apply to that
test, so they need to be run then.

The only exception is the STACK option, because of the way the test is run,
this is the only option that really can't happen inside a .TEST directive,
but it's currently a warning, not an error.

Does any of that make sense?