Reworking the warning supressing mechanism.
I now want to enable and disable warning during runtime, so the current
method won't cut it.  Since each warning has a tag with four digits, there
will never be more than 10,000 warnings.  That works out to 1,250 bytes
worth of bits, which is fine.  I can set aside 1,250 bytes to use a bitmask
to enable/disable warnings.  This work adds the support for such without
enabling its use.  This is to check in some work in progress.