New method of surpressing warnings now works.
I had to use an array of 'unsigned char' instead of 'bool' because of C
semantics (bool must be 0 or 1).  Also, I lost the ability to specify
multiple warnings with a single '-n' option, but that can be fixed.  I want
to get the dynamic enabling/disabling working first, as a proof-of-concept.