Add .FLOAT and .FLOATD directives.
For the non-rsdos backends, it will format the numbers as IEEE-754 floating
point numbers; .FLOAT is single and .FLOATD is double.  For the rsdos
backend, the .FLOAT directive will format it as a 5-byte packed floating
point number; .FLOATD will make a 6-byte unpacked format.  Expressions are
NOT supported at this time.

It is also a bit inaccurate at the larger exponents, but I think that is to
be expected.