Refactor the .FLOAT and .FLOATD directives.
We pass this on to the backend, like a lot of other directives.  Most will
use the fdefault_float() version, but the RSDOS backend will use its own
since it supports a different floating point format than IEEE-754.

This also removes some endian-specific code (sigh---perhaps Rob Pike was
correct about this) and it simplifies the code quite a bit, even if there is
a bit of duplication going on.